requirements.txt
numpy==1.16.4
matplotlib==2.2.3
Keras==2.2.2
tensorflow==1.12.0
Pillow==6.1.0
PyQt5==5.13.0
